refactor: ApiInfo::call uses rpc::Client under the hood
