+++ src/rpc_client/mpool_ops.rs
@@ -4,17 +4,13 @@
 use crate::{
     message::SignedMessage,
     rpc::{mpool_api::*, types::MessageSendSpec, RpcMethod},
-    shim::{address::Address, message::Message},
+    shim::message::Message,
 };
 use cid::Cid;
 
 use super::{ApiInfo, JsonRpcError, RpcRequest};
 
 impl ApiInfo {
-    pub fn mpool_get_nonce_req(addr: Address) -> RpcRequest<u64> {
-        RpcRequest::new(MpoolGetNonce::NAME, (addr,))
-    }
-
     pub async fn mpool_push_message(
         &self,
         message: Message,
